A basic dice game against a dragon built with JS.
This is a very basic dice game in which you are given a choice to
either fight a dragon, or run away. If you choose to fight, you have to
roll at least 8 out of 12 to beat the beast, otherwise you lose. The
game only uses HTML,CSS, and JavaScript.